Move NodeVisitors to own namespace
