+++ src/NodeVisitors/SymbolFinder.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace LanguageServer;
+namespace LanguageServer\NodeVisitors;
 
 use PhpParser\{NodeVisitorAbstract, Node};
